# target/classes/static/openapi/policies-api.yaml
openapi: 3.0.3
info:
  title: Policy Overview Dashboard API
  description: >-
    BFF (Backend-for-Frontend) API for the APAC Policy Overview Dashboard.
    Aggregates and transforms policy data from the core policy database into a
    frontend-friendly, paginated format for the Angular Material table.

    This document is the source of truth (contract-first).
  version: 2.0.0
  contact:
    name: APAC Operations Engineering
    email: apac-ops-engineering@chubb.com

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Policies

paths:
  /api/v1/policies:
    get:
      operationId: getPolicies
      summary: List policies
      description: >-
        Returns a paginated list of policies for the dashboard, transformed
        into a frontend-friendly format (display names, expiry indicator).
        Supports pagination, sorting, filtering and free-text search.
      tags:
        - Policies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          required: false
          description: Filter by policy status.
          schema:
            $ref: '#/components/schemas/PolicyStatusFilter'
        - name: lineOfBusiness
          in: query
          required: false
          description: Filter by line of business.
          schema:
            $ref: '#/components/schemas/LineOfBusinessFilter'
        - name: region
          in: query
          required: false
          description: Filter by region display name (exact match).
          schema:
            type: string
            example: Singapore
        - name: effectiveDateFrom
          in: query
          required: false
          description: Include policies effective on or after this date (inclusive).
          schema:
            type: string
            format: date
            example: 2025-01-01
        - name: effectiveDateTo
          in: query
          required: false
          description: Include policies effective on or before this date (inclusive).
          schema:
            type: string
            format: date
            example: 2025-12-31
        - name: search
          in: query
          required: false
          description: >-
            Free-text search (case-insensitive, substring) across policy
            number, policyholder name and underwriter.
          schema:
            type: string
            example: tan
      responses:
        '200':
          description: A page of policies.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPolicyResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/policies/{id}:
    get:
      operationId: getPolicyById
      summary: Get a policy by id
      description: Returns a single policy by its unique identifier.
      tags:
        - Policies
      parameters:
        - name: id
          in: path
          required: true
          description: Unique policy identifier.
          schema:
            type: string
            format: uuid
            example: 11111111-1111-1111-1111-111111111111
      responses:
        '200':
          description: The requested policy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicySummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: No policy exists with the supplied id.
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/policies/flag:
    patch:
      operationId: flagPoliciesForReview
      summary: Bulk flag policies for review
      description: >-
        Marks the supplied policies as flagged for review. Accepts an array of
        policy ids and returns no content.
      tags:
        - Policies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlagPolicyRequest'
      responses:
        '204':
          description: The policies were flagged for review.
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /api/v1/policies/summary:
    get:
      operationId: getPolicyStatistics
      summary: Aggregated policy statistics
      description: >-
        Returns aggregated dashboard statistics: counts by status, total
        premium by line of business, and the count of policies expiring soon
        (within 30 days).
      tags:
        - Policies
      responses:
        '200':
          description: Aggregated statistics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyStatisticsResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

components:

  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        format: int32
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Maximum number of records per page.
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 10
    SortParam:
      name: sort
      in: query
      required: false
      description: >-
        Sort directive in the form `property,(asc|desc)`. May be repeated to
        sort by multiple properties, e.g. `sort=premiumAmount,desc`.
      schema:
        type: array
        items:
          type: string
        example:
          - premiumAmount,desc

  responses:
    BadRequest:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: >-
        The policy service is temporarily unavailable (e.g. the policy
        database is unreachable).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:

    PolicyStatusFilter:
      type: string
      description: Policy status filter value (stored enum constant).
      enum:
        - ACTIVE
        - EXPIRED
        - PENDING
        - CANCELLED
      example: ACTIVE

    PolicyStatusDisplay:
      type: string
      description: Human-readable policy status display value.
      enum:
        - Active
        - Expired
        - Pending
        - Cancelled
      example: Active

    LineOfBusinessFilter:
      type: string
      description: Line of business filter value (stored enum constant).
      enum:
        - PROPERTY
        - CASUALTY
        - ACCIDENT_AND_HEALTH
        - MARINE
      example: MARINE

    LineOfBusinessDisplay:
      type: string
      description: Human-readable line of business display value.
      enum:
        - Property
        - Casualty
        - A&H
        - Marine
      example: Marine

    Money:
      type: object
      description: A monetary amount together with its ISO 4217 currency code.
      required:
        - amount
        - currency
      properties:
        amount:
          type: number
          format: decimal
          description: The monetary value.
          example: 12500.00
        currency:
          type: string
          description: ISO 4217 currency code.
          minLength: 3
          maxLength: 3
          example: SGD

    PolicySummary:
      type: object
      description: A single policy as presented on the dashboard list.
      required:
        - policyNumber
        - holderName
        - region
        - status
        - premium
        - startDate
        - endDate
        - isExpiringSoon
      properties:
        policyNumber:
          type: string
          description: Unique policy identifier.
          example: POL-000001
        holderName:
          type: string
          description: Full name of the policy holder.
          example: Wei Lin Tan
        region:
          type: string
          description: Region display name.
          example: Singapore
        status:
          $ref: '#/components/schemas/PolicyStatusDisplay'
        premium:
          $ref: '#/components/schemas/Money'
        startDate:
          type: string
          format: date
          description: Policy effective date (ISO 8601, yyyy-MM-dd).
          example: 2025-07-01
        endDate:
          type: string
          format: date
          description: Policy expiry date (ISO 8601, yyyy-MM-dd).
          example: 2026-07-01
        isExpiringSoon:
          type: boolean
          description: True when the expiry date is within 30 days from today.
          example: true

    PagedPolicyResponse:
      type: object
      description: >-
        Paginated list of policies, shaped for server-side pagination in an
        Angular Material table.
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          description: The policies on the current page.
          items:
            $ref: '#/components/schemas/PolicySummary'
        page:
          type: integer
          format: int32
          description: Zero-based index of the current page.
          minimum: 0
          example: 0
        size:
          type: integer
          format: int32
          description: Maximum number of records per page.
          minimum: 1
          example: 10
        totalElements:
          type: integer
          format: int64
          description: Total number of policies across all pages.
          minimum: 0
          example: 10
        totalPages:
          type: integer
          format: int32
          description: Total number of pages available.
          minimum: 0
          example: 1

    FlagPolicyRequest:
      type: object
      description: Bulk flag-for-review request.
      required:
        - policyIds
      properties:
        policyIds:
          type: array
          description: Ids of the policies to flag for review.
          minItems: 1
          items:
            type: string
            format: uuid
          example:
            - 11111111-1111-1111-1111-111111111111
            - 22222222-2222-2222-2222-222222222222

    PolicyStatisticsResponse:
      type: object
      description: Aggregated dashboard statistics.
      required:
        - countsByStatus
        - totalPremiumByLineOfBusiness
        - expiringSoonCount
      properties:
        countsByStatus:
          type: object
          description: >-
            Policy counts keyed by status display name (Active, Expired,
            Pending, Cancelled).
          additionalProperties:
            type: integer
            format: int64
          example:
            Active: 6
            Expired: 1
            Pending: 2
            Cancelled: 1
        totalPremiumByLineOfBusiness:
          type: object
          description: >-
            Total premium keyed by line of business display name. Note: amounts
            are summed across the original policy currencies and are not
            currency-normalised.
          additionalProperties:
            type: number
            format: decimal
          example:
            Property: 2940500.00
            Marine: 1765000.00
            Casualty: 194500.00
            A&H: 70200.00
        expiringSoonCount:
          type: integer
          format: int64
          description: Number of policies expiring within 30 days from today.
          minimum: 0
          example: 4

    ErrorResponse:
      type: object
      description: >-
        Standard error payload. Never exposes internal stack traces.
      required:
        - timestamp
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred (ISO 8601).
          example: 2026-06-08T07:48:55Z
        status:
          type: integer
          format: int32
          description: HTTP status code.
          example: 503
        error:
          type: string
          description: Short, human-readable error reason.
          example: Service Unavailable
        message:
          type: string
          description: Readable, non-sensitive description of the problem.
          example: The policy service is temporarily unavailable. Please try again later.
        path:
          type: string
          description: Request path that produced the error.
          example: /api/v1/policies